add class FileService fix bugs
