--- src/main/java/org/example/exception/ExceptionLackOfVisit.java
@@ -1,8 +1,0 @@
-package org.example.exception;
-
-public class ExceptionLackOfVisit extends RuntimeException {
-
-    public ExceptionLackOfVisit(String message) {
-        super(message);
-    }
-}
