+++ src/main/java/org/example/model/Doctor.java
@@ -1,7 +1,7 @@
 package org.example.model;
 
 public class Doctor extends User {
-    private Specialization specialization;
+    private final Specialization specialization;
 
     public Doctor(Integer id, String name, String surname, String login, String password, Specialization specialization) {
         super(id, name, surname, login, password);
