+++ src/main/java/org/example/model/User.java
@@ -1,12 +1,11 @@
 package org.example.model;
 
 public abstract class User {
-    private Integer id;
-    private String name;
-    private String surname;
-    private String login;
-
-    private String password;
+    private final Integer id;
+    private final String name;
+    private final String surname;
+    private final String login;
+    private final String password;
 
     public User(Integer id, String name, String surname, String login, String password) {
         this.id = id;
