+++ src/main/java/org/example/service/FileService.java
@@ -1,0 +1,16 @@
+package org.example.service;
+
+import org.example.model.Specialization;
+import org.example.model.User;
+import org.example.model.Visit;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Map;
+
+public interface FileService {
+    void writeToFileUsers(List<User> users);
+    void writeToFileVisits(Map<LocalDate, List<Visit>> listVisits);
+    List<User> getUsersFromFile();
+    Map<LocalDate, List<Visit>> getVisitsFromFile();
+}
