+++ src/main/java/org/example/service/LoginService.java
@@ -5,7 +5,7 @@ import org.example.model.User;
 import java.util.List;
 
 public interface LoginService {
-    void fakeUsers();
+//    void fakeUsers();
     User login (String login,String password);
     List<User> getUsers();
     void setUsers(List<User> users);
