+++ src/main/java/org/example/service/LoginServiceImpl.java
@@ -1,22 +1,26 @@
 package org.example.service;
 
-import org.example.exception.ExceptionLackOfVisit;
 import org.example.model.Doctor;
 import org.example.model.Patient;
 import org.example.model.Specialization;
 import org.example.model.User;
 
+import java.lang.module.FindException;
 import java.util.ArrayList;
 import java.util.List;
 
 public class LoginServiceImpl implements LoginService {
     public List<User> users = new ArrayList<>();
 
-    @Override
-    public void fakeUsers() {
-        users.add(new Patient(1, "Patient", "Patient", "patient1", "111", "322-322-322"));
-        users.add(new Doctor(2, "Doctor", "Doctor", "doctor1", "222", Specialization.CARDIOLOGIST));
-    }
+//    @Override
+//    public void fakeUsers() {
+//        users.add(new Patient(1, "Mateusz", "Chory", "patient1", "111", "322-322-322"));
+//        users.add(new Doctor(2, "Agnieszka", "Pierwsza", "doctor1", "222", Specialization.INTERN));
+//        users.add(new Doctor(3, "Adam", "Adamowicz", "doctor2", "222", Specialization.CARDIOLOGIST));
+//        users.add(new Doctor(4, "Piotr", "Slipy", "doctor3", "222", Specialization.OCULIST));
+//        users.add(new Doctor(5, "Dorota", "Pryszcz", "doctor4", "222", Specialization.DERMATOLOGIST));
+//        users.add(new Doctor(6, "Marta", "Kwiatek", "doctor5", "222", Specialization.ALLERGIST));
+//    }
 
     @Override
     public User login(String login, String password) {
@@ -25,8 +29,12 @@ public class LoginServiceImpl implements LoginService {
                 return user;
             }
         }
-        throw new ExceptionLackOfVisit("Not found User");
+        throw new FindException("Not found User");
     }
+
+//    public void addUser(Patient patient){
+//        users.add(patient);
+//    }
 
     @Override
     public List<User> getUsers() {
