+++ users.txt
@@ -1,2 +1,6 @@
-1;PatientName;PatientSurname;patient1;222;-;322-322-322
-2;DoctorName;DoctorSurname;doctor1;111;CARDIOLOGIST;-
+1;Mateusz;Chory;patient1;111;-;322-322-322
+2;Agnieszka;Pierwsza;doctor1;222;INTERN;-
+3;Adam;Adamowicz;doctor2;222;CARDIOLOGIST;-
+4;Piotr;Slipy;doctor3;222;OCULIST;-
+5;Dorota;Pryszcz;doctor4;222;DERMATOLOGIST;-
+6;Marta;Kwiatek;doctor5;222;ALLERGIST;-
