+++ visits.txt
@@ -1,4 +1,13 @@
 2024-03-01;1;09:00;2;-
-2024-03-01;2;09:30;2;1
 2024-03-01;2;09:30;2;-
-2024-03-01;3;09:30;2;-
+2024-03-01;3;10:00;3;-
+2024-03-01;4;10:30;3;-
+2024-03-01;5;11:00;4;-
+2024-03-01;6;11:30;4;-
+2024-03-01;7;12:00;5;-
+2024-03-01;8;12:30;5;-
+2024-03-01;9;13:00;6;1
+2024-03-01;9;13:00;6;-
+2024-03-01;10;13:00;6;-
+2024-03-01;11;13:00;6;-
+2024-03-01;12;13:30;6;-
